basic api and remote models di theme fixes
