+++ app/src/main/java/com/example/coinapp/MainActivity.kt
@@ -11,21 +11,30 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.coinapp.core.theme.theme.AppTheme
+import com.example.coinapp.core.di.mainModule
+import com.example.coinapp.core.theme.AppTheme
+import org.koin.compose.KoinApplication
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         enableEdgeToEdge()
         setContent {
-            AppTheme {
-                Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
-                    Greeting(
-                        name = BuildConfig.COIN_API_KEY,
-                        modifier = Modifier.padding(innerPadding)
-                    )
+            KoinApplication(application = {
+                modules(
+                    mainModule,
+                )
+            }) {
+                AppTheme {
+                    Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
+                        Greeting(
+                            name = BuildConfig.COIN_API_KEY,
+                            modifier = Modifier.padding(innerPadding)
+                        )
+                    }
                 }
             }
+
         }
     }
 }
