+++ app/src/main/java/com/example/coinapp/core/theme/Color.kt
@@ -1,0 +1,74 @@
+package com.example.coinapp.core.theme
+import androidx.compose.ui.graphics.Color
+
+val primaryLight = Color(0xFF9D1D00)
+val onPrimaryLight = Color(0xFFFFFFFF)
+val primaryContainerLight = Color(0xFFD93814)
+val onPrimaryContainerLight = Color(0xFFFFFFFF)
+val secondaryLight = Color(0xFFA03F2A)
+val onSecondaryLight = Color(0xFFFFFFFF)
+val secondaryContainerLight = Color(0xFFFF967E)
+val onSecondaryContainerLight = Color(0xFF520A00)
+val tertiaryLight = Color(0xFF694B00)
+val onTertiaryLight = Color(0xFFFFFFFF)
+val tertiaryContainerLight = Color(0xFF976D00)
+val onTertiaryContainerLight = Color(0xFFFFFFFF)
+val errorLight = Color(0xFFBA1A1A)
+val onErrorLight = Color(0xFFFFFFFF)
+val errorContainerLight = Color(0xFFFFDAD6)
+val onErrorContainerLight = Color(0xFF410002)
+val backgroundLight = Color(0xFFFFF8F6)
+val onBackgroundLight = Color(0xFF271814)
+val surfaceLight = Color(0xFFFFF8F6)
+val onSurfaceLight = Color(0xFF271814)
+val surfaceVariantLight = Color(0xFFFFDAD3)
+val onSurfaceVariantLight = Color(0xFF5B403A)
+val outlineLight = Color(0xFF907068)
+val outlineVariantLight = Color(0xFFE5BEB5)
+val scrimLight = Color(0xFF000000)
+val inverseSurfaceLight = Color(0xFF3E2C28)
+val inverseOnSurfaceLight = Color(0xFFFFEDE9)
+val inversePrimaryLight = Color(0xFFFFB4A4)
+val surfaceDimLight = Color(0xFFF1D4CD)
+val surfaceBrightLight = Color(0xFFFFF8F6)
+val surfaceContainerLowestLight = Color(0xFFFFFFFF)
+val surfaceContainerLowLight = Color(0xFFFFF0ED)
+val surfaceContainerLight = Color(0xFFFFE9E5)
+val surfaceContainerHighLight = Color(0xFFFFE2DC)
+val surfaceContainerHighestLight = Color(0xFFFADCD5)
+
+val primaryDark = Color(0xFFFFB4A4)
+val onPrimaryDark = Color(0xFF630E00)
+val primaryContainerDark = Color(0xFFD93814)
+val onPrimaryContainerDark = Color(0xFFFFFFFF)
+val secondaryDark = Color(0xFFFFB4A4)
+val onSecondaryDark = Color(0xFF611202)
+val secondaryContainerDark = Color(0xFF741F0D)
+val onSecondaryContainerDark = Color(0xFFFFC4B7)
+val tertiaryDark = Color(0xFFF5BE4B)
+val onTertiaryDark = Color(0xFF412D00)
+val tertiaryContainerDark = Color(0xFF976D00)
+val onTertiaryContainerDark = Color(0xFFFFFFFF)
+val errorDark = Color(0xFFFFB4AB)
+val onErrorDark = Color(0xFF690005)
+val errorContainerDark = Color(0xFF93000A)
+val onErrorContainerDark = Color(0xFFFFDAD6)
+val backgroundDark = Color(0xFF1E100C)
+val onBackgroundDark = Color(0xFFFADCD5)
+val surfaceDark = Color(0xFF1E100C)
+val onSurfaceDark = Color(0xFFFADCD5)
+val surfaceVariantDark = Color(0xFF5B403A)
+val onSurfaceVariantDark = Color(0xFFE5BEB5)
+val outlineDark = Color(0xFFAB8981)
+val outlineVariantDark = Color(0xFF5B403A)
+val scrimDark = Color(0xFF000000)
+val inverseSurfaceDark = Color(0xFFFADCD5)
+val inverseOnSurfaceDark = Color(0xFF3E2C28)
+val inversePrimaryDark = Color(0xFFB72300)
+val surfaceDimDark = Color(0xFF1E100C)
+val surfaceBrightDark = Color(0xFF483530)
+val surfaceContainerLowestDark = Color(0xFF180A08)
+val surfaceContainerLowDark = Color(0xFF271814)
+val surfaceContainerDark = Color(0xFF2C1C18)
+val surfaceContainerHighDark = Color(0xFF372622)
+val surfaceContainerHighestDark = Color(0xFF43302C)
