+++ app/src/main/java/com/example/coinapp/core/theme/Type.kt
@@ -1,0 +1,5 @@
+package com.example.coinapp.core.theme
+
+import androidx.compose.material3.Typography
+
+val AppTypography = Typography()
