--- app/src/main/java/com/example/coinapp/core/theme/theme/Color.kt
@@ -1,225 +1,0 @@
-package com.example.coinapp.core.theme.theme
-import androidx.compose.ui.graphics.Color
-
-val primaryLight = Color(0xFF9D1D00)
-val onPrimaryLight = Color(0xFFFFFFFF)
-val primaryContainerLight = Color(0xFFD93814)
-val onPrimaryContainerLight = Color(0xFFFFFFFF)
-val secondaryLight = Color(0xFFA03F2A)
-val onSecondaryLight = Color(0xFFFFFFFF)
-val secondaryContainerLight = Color(0xFFFF967E)
-val onSecondaryContainerLight = Color(0xFF520A00)
-val tertiaryLight = Color(0xFF694B00)
-val onTertiaryLight = Color(0xFFFFFFFF)
-val tertiaryContainerLight = Color(0xFF976D00)
-val onTertiaryContainerLight = Color(0xFFFFFFFF)
-val errorLight = Color(0xFFBA1A1A)
-val onErrorLight = Color(0xFFFFFFFF)
-val errorContainerLight = Color(0xFFFFDAD6)
-val onErrorContainerLight = Color(0xFF410002)
-val backgroundLight = Color(0xFFFFF8F6)
-val onBackgroundLight = Color(0xFF271814)
-val surfaceLight = Color(0xFFFFF8F6)
-val onSurfaceLight = Color(0xFF271814)
-val surfaceVariantLight = Color(0xFFFFDAD3)
-val onSurfaceVariantLight = Color(0xFF5B403A)
-val outlineLight = Color(0xFF907068)
-val outlineVariantLight = Color(0xFFE5BEB5)
-val scrimLight = Color(0xFF000000)
-val inverseSurfaceLight = Color(0xFF3E2C28)
-val inverseOnSurfaceLight = Color(0xFFFFEDE9)
-val inversePrimaryLight = Color(0xFFFFB4A4)
-val surfaceDimLight = Color(0xFFF1D4CD)
-val surfaceBrightLight = Color(0xFFFFF8F6)
-val surfaceContainerLowestLight = Color(0xFFFFFFFF)
-val surfaceContainerLowLight = Color(0xFFFFF0ED)
-val surfaceContainerLight = Color(0xFFFFE9E5)
-val surfaceContainerHighLight = Color(0xFFFFE2DC)
-val surfaceContainerHighestLight = Color(0xFFFADCD5)
-
-val primaryLightMediumContrast = Color(0xFF851700)
-val onPrimaryLightMediumContrast = Color(0xFFFFFFFF)
-val primaryContainerLightMediumContrast = Color(0xFFD93814)
-val onPrimaryContainerLightMediumContrast = Color(0xFFFFFFFF)
-val secondaryLightMediumContrast = Color(0xFF7B2512)
-val onSecondaryLightMediumContrast = Color(0xFFFFFFFF)
-val secondaryContainerLightMediumContrast = Color(0xFFBD553D)
-val onSecondaryContainerLightMediumContrast = Color(0xFFFFFFFF)
-val tertiaryLightMediumContrast = Color(0xFF583E00)
-val onTertiaryLightMediumContrast = Color(0xFFFFFFFF)
-val tertiaryContainerLightMediumContrast = Color(0xFF976D00)
-val onTertiaryContainerLightMediumContrast = Color(0xFFFFFFFF)
-val errorLightMediumContrast = Color(0xFF8C0009)
-val onErrorLightMediumContrast = Color(0xFFFFFFFF)
-val errorContainerLightMediumContrast = Color(0xFFDA342E)
-val onErrorContainerLightMediumContrast = Color(0xFFFFFFFF)
-val backgroundLightMediumContrast = Color(0xFFFFF8F6)
-val onBackgroundLightMediumContrast = Color(0xFF271814)
-val surfaceLightMediumContrast = Color(0xFFFFF8F6)
-val onSurfaceLightMediumContrast = Color(0xFF271814)
-val surfaceVariantLightMediumContrast = Color(0xFFFFDAD3)
-val onSurfaceVariantLightMediumContrast = Color(0xFF573C36)
-val outlineLightMediumContrast = Color(0xFF765851)
-val outlineVariantLightMediumContrast = Color(0xFF94736C)
-val scrimLightMediumContrast = Color(0xFF000000)
-val inverseSurfaceLightMediumContrast = Color(0xFF3E2C28)
-val inverseOnSurfaceLightMediumContrast = Color(0xFFFFEDE9)
-val inversePrimaryLightMediumContrast = Color(0xFFFFB4A4)
-val surfaceDimLightMediumContrast = Color(0xFFF1D4CD)
-val surfaceBrightLightMediumContrast = Color(0xFFFFF8F6)
-val surfaceContainerLowestLightMediumContrast = Color(0xFFFFFFFF)
-val surfaceContainerLowLightMediumContrast = Color(0xFFFFF0ED)
-val surfaceContainerLightMediumContrast = Color(0xFFFFE9E5)
-val surfaceContainerHighLightMediumContrast = Color(0xFFFFE2DC)
-val surfaceContainerHighestLightMediumContrast = Color(0xFFFADCD5)
-
-val primaryLightHighContrast = Color(0xFF490800)
-val onPrimaryLightHighContrast = Color(0xFFFFFFFF)
-val primaryContainerLightHighContrast = Color(0xFF851700)
-val onPrimaryContainerLightHighContrast = Color(0xFFFFFFFF)
-val secondaryLightHighContrast = Color(0xFF490800)
-val onSecondaryLightHighContrast = Color(0xFFFFFFFF)
-val secondaryContainerLightHighContrast = Color(0xFF7B2512)
-val onSecondaryContainerLightHighContrast = Color(0xFFFFFFFF)
-val tertiaryLightHighContrast = Color(0xFF2F1F00)
-val onTertiaryLightHighContrast = Color(0xFFFFFFFF)
-val tertiaryContainerLightHighContrast = Color(0xFF583E00)
-val onTertiaryContainerLightHighContrast = Color(0xFFFFFFFF)
-val errorLightHighContrast = Color(0xFF4E0002)
-val onErrorLightHighContrast = Color(0xFFFFFFFF)
-val errorContainerLightHighContrast = Color(0xFF8C0009)
-val onErrorContainerLightHighContrast = Color(0xFFFFFFFF)
-val backgroundLightHighContrast = Color(0xFFFFF8F6)
-val onBackgroundLightHighContrast = Color(0xFF271814)
-val surfaceLightHighContrast = Color(0xFFFFF8F6)
-val onSurfaceLightHighContrast = Color(0xFF000000)
-val surfaceVariantLightHighContrast = Color(0xFFFFDAD3)
-val onSurfaceVariantLightHighContrast = Color(0xFF351E19)
-val outlineLightHighContrast = Color(0xFF573C36)
-val outlineVariantLightHighContrast = Color(0xFF573C36)
-val scrimLightHighContrast = Color(0xFF000000)
-val inverseSurfaceLightHighContrast = Color(0xFF3E2C28)
-val inverseOnSurfaceLightHighContrast = Color(0xFFFFFFFF)
-val inversePrimaryLightHighContrast = Color(0xFFFFE7E2)
-val surfaceDimLightHighContrast = Color(0xFFF1D4CD)
-val surfaceBrightLightHighContrast = Color(0xFFFFF8F6)
-val surfaceContainerLowestLightHighContrast = Color(0xFFFFFFFF)
-val surfaceContainerLowLightHighContrast = Color(0xFFFFF0ED)
-val surfaceContainerLightHighContrast = Color(0xFFFFE9E5)
-val surfaceContainerHighLightHighContrast = Color(0xFFFFE2DC)
-val surfaceContainerHighestLightHighContrast = Color(0xFFFADCD5)
-
-val primaryDark = Color(0xFFFFB4A4)
-val onPrimaryDark = Color(0xFF630E00)
-val primaryContainerDark = Color(0xFFD93814)
-val onPrimaryContainerDark = Color(0xFFFFFFFF)
-val secondaryDark = Color(0xFFFFB4A4)
-val onSecondaryDark = Color(0xFF611202)
-val secondaryContainerDark = Color(0xFF741F0D)
-val onSecondaryContainerDark = Color(0xFFFFC4B7)
-val tertiaryDark = Color(0xFFF5BE4B)
-val onTertiaryDark = Color(0xFF412D00)
-val tertiaryContainerDark = Color(0xFF976D00)
-val onTertiaryContainerDark = Color(0xFFFFFFFF)
-val errorDark = Color(0xFFFFB4AB)
-val onErrorDark = Color(0xFF690005)
-val errorContainerDark = Color(0xFF93000A)
-val onErrorContainerDark = Color(0xFFFFDAD6)
-val backgroundDark = Color(0xFF1E100C)
-val onBackgroundDark = Color(0xFFFADCD5)
-val surfaceDark = Color(0xFF1E100C)
-val onSurfaceDark = Color(0xFFFADCD5)
-val surfaceVariantDark = Color(0xFF5B403A)
-val onSurfaceVariantDark = Color(0xFFE5BEB5)
-val outlineDark = Color(0xFFAB8981)
-val outlineVariantDark = Color(0xFF5B403A)
-val scrimDark = Color(0xFF000000)
-val inverseSurfaceDark = Color(0xFFFADCD5)
-val inverseOnSurfaceDark = Color(0xFF3E2C28)
-val inversePrimaryDark = Color(0xFFB72300)
-val surfaceDimDark = Color(0xFF1E100C)
-val surfaceBrightDark = Color(0xFF483530)
-val surfaceContainerLowestDark = Color(0xFF180A08)
-val surfaceContainerLowDark = Color(0xFF271814)
-val surfaceContainerDark = Color(0xFF2C1C18)
-val surfaceContainerHighDark = Color(0xFF372622)
-val surfaceContainerHighestDark = Color(0xFF43302C)
-
-val primaryDarkMediumContrast = Color(0xFFFFBAAB)
-val onPrimaryDarkMediumContrast = Color(0xFF340400)
-val primaryContainerDarkMediumContrast = Color(0xFFFF5632)
-val onPrimaryContainerDarkMediumContrast = Color(0xFF000000)
-val secondaryDarkMediumContrast = Color(0xFFFFBAAB)
-val onSecondaryDarkMediumContrast = Color(0xFF340400)
-val secondaryContainerDarkMediumContrast = Color(0xFFE17056)
-val onSecondaryContainerDarkMediumContrast = Color(0xFF000000)
-val tertiaryDarkMediumContrast = Color(0xFFFAC24F)
-val onTertiaryDarkMediumContrast = Color(0xFF201400)
-val tertiaryContainerDarkMediumContrast = Color(0xFFB98815)
-val onTertiaryContainerDarkMediumContrast = Color(0xFF000000)
-val errorDarkMediumContrast = Color(0xFFFFBAB1)
-val onErrorDarkMediumContrast = Color(0xFF370001)
-val errorContainerDarkMediumContrast = Color(0xFFFF5449)
-val onErrorContainerDarkMediumContrast = Color(0xFF000000)
-val backgroundDarkMediumContrast = Color(0xFF1E100C)
-val onBackgroundDarkMediumContrast = Color(0xFFFADCD5)
-val surfaceDarkMediumContrast = Color(0xFF1E100C)
-val onSurfaceDarkMediumContrast = Color(0xFFFFF9F8)
-val surfaceVariantDarkMediumContrast = Color(0xFF5B403A)
-val onSurfaceVariantDarkMediumContrast = Color(0xFFE9C2BA)
-val outlineDarkMediumContrast = Color(0xFFBF9B93)
-val outlineVariantDarkMediumContrast = Color(0xFF9D7B74)
-val scrimDarkMediumContrast = Color(0xFF000000)
-val inverseSurfaceDarkMediumContrast = Color(0xFFFADCD5)
-val inverseOnSurfaceDarkMediumContrast = Color(0xFF372622)
-val inversePrimaryDarkMediumContrast = Color(0xFF8E1900)
-val surfaceDimDarkMediumContrast = Color(0xFF1E100C)
-val surfaceBrightDarkMediumContrast = Color(0xFF483530)
-val surfaceContainerLowestDarkMediumContrast = Color(0xFF180A08)
-val surfaceContainerLowDarkMediumContrast = Color(0xFF271814)
-val surfaceContainerDarkMediumContrast = Color(0xFF2C1C18)
-val surfaceContainerHighDarkMediumContrast = Color(0xFF372622)
-val surfaceContainerHighestDarkMediumContrast = Color(0xFF43302C)
-
-val primaryDarkHighContrast = Color(0xFFFFF9F8)
-val onPrimaryDarkHighContrast = Color(0xFF000000)
-val primaryContainerDarkHighContrast = Color(0xFFFFBAAB)
-val onPrimaryContainerDarkHighContrast = Color(0xFF000000)
-val secondaryDarkHighContrast = Color(0xFFFFF9F8)
-val onSecondaryDarkHighContrast = Color(0xFF000000)
-val secondaryContainerDarkHighContrast = Color(0xFFFFBAAB)
-val onSecondaryContainerDarkHighContrast = Color(0xFF000000)
-val tertiaryDarkHighContrast = Color(0xFFFFFAF7)
-val onTertiaryDarkHighContrast = Color(0xFF000000)
-val tertiaryContainerDarkHighContrast = Color(0xFFFAC24F)
-val onTertiaryContainerDarkHighContrast = Color(0xFF000000)
-val errorDarkHighContrast = Color(0xFFFFF9F9)
-val onErrorDarkHighContrast = Color(0xFF000000)
-val errorContainerDarkHighContrast = Color(0xFFFFBAB1)
-val onErrorContainerDarkHighContrast = Color(0xFF000000)
-val backgroundDarkHighContrast = Color(0xFF1E100C)
-val onBackgroundDarkHighContrast = Color(0xFFFADCD5)
-val surfaceDarkHighContrast = Color(0xFF1E100C)
-val onSurfaceDarkHighContrast = Color(0xFFFFFFFF)
-val surfaceVariantDarkHighContrast = Color(0xFF5B403A)
-val onSurfaceVariantDarkHighContrast = Color(0xFFFFF9F8)
-val outlineDarkHighContrast = Color(0xFFE9C2BA)
-val outlineVariantDarkHighContrast = Color(0xFFE9C2BA)
-val scrimDarkHighContrast = Color(0xFF000000)
-val inverseSurfaceDarkHighContrast = Color(0xFFFADCD5)
-val inverseOnSurfaceDarkHighContrast = Color(0xFF000000)
-val inversePrimaryDarkHighContrast = Color(0xFF580B00)
-val surfaceDimDarkHighContrast = Color(0xFF1E100C)
-val surfaceBrightDarkHighContrast = Color(0xFF483530)
-val surfaceContainerLowestDarkHighContrast = Color(0xFF180A08)
-val surfaceContainerLowDarkHighContrast = Color(0xFF271814)
-val surfaceContainerDarkHighContrast = Color(0xFF2C1C18)
-val surfaceContainerHighDarkHighContrast = Color(0xFF372622)
-val surfaceContainerHighestDarkHighContrast = Color(0xFF43302C)
-
-
-
-
-
-
-
