--- app/src/main/java/com/example/coinapp/core/theme/theme/Type.kt
@@ -1,9 +1,0 @@
-package com.example.coinapp.core.theme.theme
-
-import androidx.compose.material3.Typography
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.sp
-
-val AppTypography = Typography()
