+++ secrets.properties
@@ -1,1 +1,3 @@
-COIN_API_KEY = "6b9f0349-106f-44a4-bd14-90f248045642"
+COIN_API_KEY="6b9f0349-106f-44a4-bd14-90f248045642"
+BASE_URL="https://rest.coinapi.io/v1"
+ICONS_URL="https://s3.eu-central-1.amazonaws.com/bbxt-static-icons/type-id/png_512/%s.png"
